Restore axis-ptz
